+++ packages/provider-utils/src/types/tool.test-d.ts
@@ -82,7 +82,12 @@ describe('tool type', () => {
       });
 
       expectTypeOf(aTool.toModelOutput).toMatchTypeOf<
-        ((options: { output: any }) => ToolResultOutput) | undefined
+        | ((options: {
+            toolCallId: string;
+            input: { number: number };
+            output: any;
+          }) => ToolResultOutput | PromiseLike<ToolResultOutput>)
+        | undefined
       >();
     });
 
@@ -97,7 +102,12 @@ describe('tool type', () => {
       });
 
       expectTypeOf(aTool.toModelOutput).toMatchTypeOf<
-        ((options: { output: 'test' }) => ToolResultOutput) | undefined
+        | ((options: {
+            toolCallId: string;
+            input: { number: number };
+            output: 'test';
+          }) => ToolResultOutput | PromiseLike<ToolResultOutput>)
+        | undefined
       >();
     });
 
@@ -112,7 +122,12 @@ describe('tool type', () => {
       });
 
       expectTypeOf(aTool.toModelOutput).toMatchTypeOf<
-        ((options: { output: 'test' }) => ToolResultOutput) | undefined
+        | ((options: {
+            toolCallId: string;
+            input: { number: number };
+            output: 'test';
+          }) => ToolResultOutput | PromiseLike<ToolResultOutput>)
+        | undefined
       >();
     });
   });